pyproject.toml
[tool.poetry]
name = "photo-processor"
version = "0.1.0"
description = "Photo processing toolkit"
authors = ["ashic <ashic@live.com>"]
license = "Moz-2"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fire = "^0.6.0"
pillow = "^10.4.0"
exif = "^1.6.0"
xlrd = "^2.0.1"
pandas = "^2.2.2"
openpyxl = "^3.1.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
